Split the tasks into their own file, separate from the views, and cleaned it up with some docstrings and comments.
